+++ mybatis-plus-extension/src/main/java/com/baomidou/mybatisplus/extension/plugins/PaginationInterceptor.java
@@ -93,7 +93,7 @@ public class PaginationInterceptor extends AbstractSqlParserHandler implements I
     /**
      * 单页限制 500 条，小于 0 如 -1 不受限制
      */
-    private long limit = 500L;
+    private int limit = 500;
     /**
      * 方言类型(数据库名,全小写) <br>
      * 如果用的我们支持分页的数据库但获取数据库类型不正确则可以配置该值进行校正
@@ -251,10 +251,10 @@ public class PaginationInterceptor extends AbstractSqlParserHandler implements I
         try (PreparedStatement statement = connection.prepareStatement(sql)) {
             DefaultParameterHandler parameterHandler = new MybatisDefaultParameterHandler(mappedStatement, boundSql.getParameterObject(), boundSql);
             parameterHandler.setParameters(statement);
-            long total = 0;
+            int total = 0;
             try (ResultSet resultSet = statement.executeQuery()) {
                 if (resultSet.next()) {
-                    total = resultSet.getLong(1);
+                    total = resultSet.getInt(1);
                 }
             }
             page.setTotal(total);
